+++ src/video/renderer.js
@@ -90,7 +90,7 @@ export default class Renderer {
         // create the main screen canvas
         if (platform.ejecta === true) {
             // a main canvas is already automatically created by Ejecta
-            this.canvas = document.getElementById("canvas");
+            this.canvas = globalThis.document.getElementById("canvas");
         } else if (typeof globalThis.canvas !== "undefined") {
             // a global canvas is available, e.g. webapp adapter for wechat
             this.canvas = globalThis.canvas;